backend-bulk/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0  # Async PostgreSQL driver
psycopg2-binary==2.9.9
alembic==1.12.1
supabase==2.10.0  # Working version with httpx>=0.27


# Task Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client
httpx>=0.27.0
requests==2.31.0

# PDF Processing
PyMuPDF==1.24.0
Pillow==10.2.0
numpy>=1.24.0,<2.0.0  # NumPy for OpenCV compatibility
opencv-python-headless==4.9.0.80  # OpenCV for image processing (headless for server)

# Excel/CSV Export
openpyxl==3.1.2

# Cloud Storage & Drive Integration
google-api-python-client==2.108.0  # Google Drive API
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
msal==1.26.0  # Microsoft Authentication Library for OneDrive

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

